Updated the player stats page
I refactored the achievements page to have less medals and i moved those stats over to the info page. I took out the div for the username and pfp, and moved those into the page nav bar. I also created a 3 column div of stats and added responsiveness to the nar bar. Wll finish the rest of responsiveness after lunch.